requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
arrow==1.3.0
    # via cookiecutter
binaryornot==0.4.4
    # via cookiecutter
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via binaryornot
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   cookiecutter
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   qrcode
    #   tqdm
    #   uvicorn
cookiecutter==2.6.0
    # via flet-cli
ezdxf==1.4.2
faiss-cpu==1.11.0
fastapi==0.115.13
    # via flet-web
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flet==0.28.3
    # via
    #   flet-cli
    #   flet-desktop
    #   flet-web
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
fonttools==4.58.4
    # via ezdxf
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.5
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via flet
huggingface-hub==0.33.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   cookiecutter
    #   torch
joblib==1.5.1
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   ezdxf
    #   faiss-cpu
    #   scikit-learn
    #   scipy
    #   shapely
    #   transformers
oauthlib==3.3.1
    # via flet
packaging==25.0
    # via
    #   faiss-cpu
    #   flet-cli
    #   huggingface-hub
    #   transformers
pillow==11.2.1
    # via sentence-transformers
pip==23.2.1
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyparsing==3.2.3
    # via ezdxf
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.9.0.post0
    # via arrow
python-dotenv==1.1.0
    # via uvicorn
python-slugify==8.0.4
    # via cookiecutter
pyyaml==6.0.2
    # via
    #   cookiecutter
    #   huggingface-hub
    #   transformers
    #   uvicorn
qrcode==7.4.2
    # via flet-cli
regex==2024.11.6
    # via transformers
repath==0.9.0
    # via flet
requests==2.32.4
    # via
    #   cookiecutter
    #   huggingface-hub
    #   transformers
rich==14.0.0
    # via cookiecutter
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via sentence-transformers
scipy==1.16.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.1.0
setuptools==80.9.0
    # via torch
shapely==2.1.1
six==1.17.0
    # via
    #   python-dateutil
    #   repath
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
toml==0.10.2
    # via flet-cli
torch==2.7.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.52.4
    # via sentence-transformers
types-python-dateutil==2.9.0.20250516
    # via arrow
typing-extensions==4.14.0
    # via
    #   ezdxf
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   qrcode
    #   sentence-transformers
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uv==0.7.13
uvicorn==0.34.3
    # via flet-web
watchdog==4.0.2
    # via flet-cli
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
whoosh==2.7.4
